Render at most one dialog at a time
