+++ src/components/Search.tsx
@@ -2,19 +2,10 @@
  * Copyright (c) Facebook, Inc. and its affiliates.
  */
 
-// @ts-ignore
-import {IconSearch} from 'components/Icon/IconSearch';
 import Head from 'next/head';
 import Link from 'next/link';
 import Router from 'next/router';
-import {
-  lazy,
-  Suspense,
-  useState,
-  startTransition,
-  useCallback,
-  useEffect,
-} from 'react';
+import {lazy, useCallback, useEffect} from 'react';
 import * as React from 'react';
 import {createPortal} from 'react-dom';
 import {siteConfig} from 'siteConfig';
@@ -25,7 +16,9 @@ export interface SearchProps {
   apiKey?: string;
   indexName?: string;
   searchParameters?: any;
-  renderModal?: boolean;
+  isOpen: boolean;
+  onOpen: () => void;
+  onClose: () => void;
 }
 
 function Hit({hit, children}: any) {
@@ -33,18 +26,6 @@ function Hit({hit, children}: any) {
     <Link href={hit.url.replace()}>
       <a>{children}</a>
     </Link>
-  );
-}
-
-function Kbd(props: {children?: React.ReactNode; wide?: boolean}) {
-  const {wide, ...rest} = props;
-  const width = wide ? 'w-10' : 'w-5';
-
-  return (
-    <kbd
-      className={`${width} h-5 border border-transparent mr-1 bg-wash dark:bg-wash-dark text-gray-30 align-middle p-0 inline-flex justify-center items-center text-xs text-center rounded-md`}
-      {...rest}
-    />
   );
 }
 
@@ -107,30 +88,21 @@ const options = {
 };
 
 const DocSearchModal: any = lazy(() =>
+  // @ts-ignore
   import('@docsearch/react/modal').then((mod) => ({
     default: mod.DocSearchModal,
   }))
 );
 
 export function Search({
+  isOpen,
+  onOpen,
+  onClose,
   searchParameters = {
     hitsPerPage: 5,
   },
 }: SearchProps) {
-  const [isShowing, setIsShowing] = useState(false);
-
-  const onOpen = useCallback(() => {
-    startTransition(() => {
-      setIsShowing(true);
-    });
-  }, []);
-
-  const onClose = useCallback(() => {
-    setIsShowing(false);
-  }, []);
-
-  useDocSearchKeyboardEvents({isOpen: isShowing, onOpen, onClose});
-
+  useDocSearchKeyboardEvents({isOpen, onOpen, onClose});
   return (
     <>
       <Head>
@@ -139,33 +111,7 @@ export function Search({
           href={`https://${options.appId}-dsn.algolia.net`}
         />
       </Head>
-
-      <button
-        aria-label="Search"
-        type="button"
-        className="active:scale-95 transition-transform flex md:hidden w-12 h-12 rounded-full items-center justify-center hover:bg-secondary-button hover:dark:bg-secondary-button-dark outline-link"
-        onClick={onOpen}>
-        <IconSearch className="align-middle w-5 h-5" />
-      </button>
-
-      <button
-        type="button"
-        className={cn(
-          'hidden md:flex 3xl:w-[56rem] 3xl:mx-0 relative pl-4 pr-1 py-1 h-10 bg-gray-30/20 dark:bg-gray-40/20 outline-none  focus:outline-link betterhover:hover:bg-opacity-80 pointer items-center text-left w-full text-gray-30 rounded-full align-middle text-base'
-        )}
-        onClick={onOpen}>
-        <IconSearch className="mr-3 align-middle text-gray-30 shrink-0 group-betterhover:hover:text-gray-70" />
-        Search
-        <span className="ml-auto hidden sm:flex item-center mr-1">
-          <Kbd data-platform="mac">⌘</Kbd>
-          <Kbd data-platform="win" wide>
-            Ctrl
-          </Kbd>
-          <Kbd>K</Kbd>
-        </span>
-      </button>
-
-      {isShowing &&
+      {isOpen &&
         createPortal(
           <DocSearchModal
             {...options}
